chore: format files module wise
