+++ client/src/components/place/PlaceCard.tsx
